updated EquipmentHandler script: merged ItemEquipped/Unequipped events into 1 OnEquippedItemChanges with added bool arg wasEquipped.
updated StateHandler, InventoryHandler, InventorySlotUi scripts:
all use new OnEquippedItemChanges event. updated internal logic for each function to be simpler.

updated NpcStateMachine script:
added logic to manage new UseConsumableState and updated some existing ones.

updated NpcBeliefs script:
added beliefs for using consuamble types.
added logic to cache and track equipment changes for consumables and to cache found usable ones to save performance and making checking/reading info easier.

fixed bug in NpcPerception, npc no longer investigate damage done by themselves (and inturn when debug damaging them for testing)

added new NpcUseConsumableState script:
simple script that when done could play an animation and wait on use consumable time. atm hard coded to wait 2.5s to simulate consumable use time before then using said consumable.

TODO: add logic so npcs heal/drink/eat till there stat is full while possible. update UseConsumable method with bool arg to not use consumable. (let npc have unlimited water/food as an option)

+++ Assets/Scripts/ItemSystem/Definitions/ArmourDefinition.cs
@@ -14,7 +14,7 @@ public class ArmourDefinition : ItemDefinition
 	[Tooltip("% Damage reduction")]
 	[Range(0f, 1f)]
 	[SerializeField] private float protectionProvided;
-	[SerializeField] private float inventorySlotsProvided;
+	[SerializeField] private int inventorySlotsProvided;
 	#endregion
 
 	//add fields for ui icons, 3d prefab models etc, sfx/vfx specific for armour etc...
@@ -24,6 +24,6 @@ public class ArmourDefinition : ItemDefinition
 	#region readoly properties
 	public ArmourSlotType ArmourSlot => armourType;
 	public float ProtectionProvided => protectionProvided;
-	public float InventorySlotsProvided => inventorySlotsProvided;
+	public int InventorySlotsProvided => inventorySlotsProvided;
 	#endregion
 }
